+++ packages/route-pattern/src/lib/href.test.ts
@@ -1,7 +1,7 @@
 import * as assert from 'node:assert/strict'
 import { describe, it } from 'node:test'
 
-import { createHrefBuilder } from './href.ts'
+import { MissingParamError, createHrefBuilder } from './href.ts'
 
 describe('href', () => {
   it('uses a default protocol', () => {
@@ -25,49 +25,54 @@ describe('href', () => {
     assert.equal(href('*/files/*.jpg', { '*': 'cat/dog' }), '/cat/dog/files/cat/dog.jpg')
   })
 
-  it('throws when using optional in href()', () => {
-    let href = createHrefBuilder()
-    assert.throws(
-      () => href('products/:id(.:ext)', { id: '1' }),
-      new Error('Cannot use pattern with optional in href()'),
-    )
-  })
+  it('fills in params', () => {
+    let href = createHrefBuilder({ host: 'remix.run' })
+    assert.equal(href('products/:id', { id: '1' }), 'https://remix.run/products/1')
+    assert.equal(href('products/:id', { id: 1 }), 'https://remix.run/products/1')
 
-  it('throws when using enum in href()', () => {
-    let href = createHrefBuilder()
-    assert.throws(
-      () => href('products/:id.{jpg,png}', { id: 'blah' }),
-      new Error('Cannot use pattern with enum in href()'),
+    assert.equal(
+      href('images/*path.png', { path: 'images/hero' }),
+      'https://remix.run/images/images/hero.png',
     )
+
+    assert.equal(
+      href('images/*.png', { '*': 'images/hero' }),
+      'https://remix.run/images/images/hero.png',
+    )
+
+    // Use the first member of an enum
+    assert.equal(
+      href('images/:id.{jpg,png}', { id: 'remix' }),
+      'https://remix.run/images/remix.jpg',
+    )
+
+    // Include optionals by default
+    assert.equal(href('products(.md)'), 'https://remix.run/products.md')
+
+    // Omit optionals with undefined/missing params
+    assert.equal(href('products/:id(.:ext)', { id: '1' }), 'https://remix.run/products/1')
+    assert.equal(href('products(/:id)', {}), 'https://remix.run/products')
+    assert.equal(href('products(/:id)', null), 'https://remix.run/products')
   })
 
   it('enforces type safety', () => {
     let href = createHrefBuilder<'products(/:id)'>()
-    // @ts-expect-error invalid variant
+    // @ts-expect-error invalid pattern
     assert.equal(href('does-not-exist'), '/does-not-exist')
-    // @ts-expect-error extraneous param arg
-    assert.equal(href('://remix.run/about', {}), 'https://remix.run/about')
-    // @ts-expect-error missing param arg
-    assert.throws(() => href('products/:id', {}), new Error('Missing required parameter: id'))
-    // @ts-expect-error missing param keys
-    assert.throws(() => href('products/:id', {}), new Error('Missing required parameter: id'))
-    // @ts-expect-error non-string param values
-    assert.equal(href('products/:id', { id: 1 }), '/products/1')
+    // @ts-expect-error null param value
+    assert.equal(href('products(/:id)', { id: null }), '/products')
+    // @ts-expect-error undefined param value
+    assert.equal(href('products(/:id)', { id: undefined }), '/products')
   })
 
   it('throws when required params are missing', () => {
     let href = createHrefBuilder()
     // @ts-expect-error missing param keys
-    assert.throws(() => href('products/:id', {}), new Error('Missing required parameter: id'))
+    assert.throws(() => href('products/:id', {}), new MissingParamError('id'))
     // @ts-expect-error missing param keys
-    assert.throws(() => href('products/*path', {}), new Error('Missing required parameter: path'))
+    assert.throws(() => href('products/*path', {}), new MissingParamError('path'))
     // @ts-expect-error missing param keys
-    assert.throws(() => href('products/*', {}), new Error('Missing required parameter: *'))
-  })
-
-  it('fills in params', () => {
-    let href = createHrefBuilder({ host: 'remix.run' })
-    assert.equal(href('products/:id', { id: '1' }), 'https://remix.run/products/1')
+    assert.throws(() => href('products/*', {}), new MissingParamError('*'))
   })
 
   it('fills in search params', () => {